fixes from messy merge
