+++ tests/utils/test_utilities.py
@@ -315,3 +315,26 @@ def test_hist():
 
     with pytest.raises(XGIError):
         xgi.hist(vals, bins="ten")
+
+
+def test_binomial_sequence():
+
+    # ask for more ones than available characters
+    output = xgi.binomial_sequence(3, 2)
+    expected_output = set()
+    assert output == expected_output
+
+    # ask for an empty string
+    output = xgi.binomial_sequence(0, 0)
+    expected_output = {""}
+    assert output == expected_output
+
+    # ask for only zeros
+    output = xgi.binomial_sequence(0, 2)
+    expected_output = {"00"}
+    assert output == expected_output
+
+    # regular output
+    output = xgi.binomial_sequence(2, 4)
+    expected_output = {"1100", "1001", "0011", "1010", "0101", "0110"}
+    assert output == expected_output
